use double splat for keyword argument expansion
