+++ spec/lightspeed/client_spec.rb
@@ -2,21 +2,6 @@
 
 require 'spec_helper'
 
-class ExampleTokenHolder
-  def initialize(token: nil, refresh_token: nil)
-    @token = token
-    @refresh_token = refresh_token
-  end
-
-  def oauth_token
-    @token
-  end
-
-  def refresh_oauth_token
-    @token = @token + "_refreshed"
-  end
-end
-
 describe Lightspeed::Client do
   it "can set an oauth token holder and fetch an access token from it" do
     key = 'test'
